requirements.txt
# DeepRead AI - Dependencies

# Web UI
pywebview>=4.4

# PDF Processing
PyMuPDF>=1.23
Pillow>=10.0

# AI Integration (optional but recommended)
openai>=1.0
requests>=2.28

# Development
pytest>=7.0
pytest-cov>=4.0
